med sized refactoring out of all the old models that weren't layer based
